fix timeline page
