+++ src/components/ReviewFeed/index.ts
@@ -1,0 +1,1 @@
+export { ReviewFeed } from "./ReviewFeed";
